+++ lib/model/certificate_model.dart
@@ -22,7 +22,6 @@ List<CertificateModel> certificateList = [
     skills: 'Flutter . Dart',
     credential:  'https://www.coursera.org/account/accomplishments/certificate/HQ4LFHSF4LKZ',
   ),
-
   CertificateModel(
     name: 'Flutter Essential Training: Build for Multiple Platforms',
     organization: 'LinkedIn',
@@ -30,7 +29,6 @@ List<CertificateModel> certificateList = [
     skills: 'Flutter . iOS Development . Android Development',
     credential:  'https://www.linkedin.com/learning/certificates/450fc4e2f495726aea50a067caf586869ccf0cb92ebcc5a4c7b5648a95754a8f',
   ),
-
   CertificateModel(
     name: 'Complete Dart Learning',
     organization: 'Udemy',
@@ -38,7 +36,6 @@ List<CertificateModel> certificateList = [
     skills: 'Flutter . Dart . Programming',
     credential:  'https://www.udemy.com/certificate/UC-5b01c756-0d20-4342-94e6-9d5860d1c95e/',
   ),
-
   CertificateModel(
     name: 'Flutter REST Movie App',
     organization: 'Udemy',
@@ -46,7 +43,6 @@ List<CertificateModel> certificateList = [
     skills: 'Flutter . Rest API\'s . Cloud',
     credential:  'https://www.udemy.com/certificate/UC-22efc7ca-3df5-4f26-8025-4a1bd2672f19/',
   ),
-
   CertificateModel(
     name: 'Modularizing and Organizing Flutter Code',
     organization: 'LinkedIn',
@@ -54,9 +50,6 @@ List<CertificateModel> certificateList = [
     skills: 'Flutter . Clean Architecture',
     credential:  'https://www.linkedin.com/learning/certificates/686276fa42629d4f1291da79ea46bfde0222954b60297a2e728c770768f23407',
   ),
-
-
-
   CertificateModel(
     name: 'Powering Your App with Live Web Data',
     organization: 'LinkedIn',
@@ -64,9 +57,6 @@ List<CertificateModel> certificateList = [
     skills: 'Flutter . Dart . Firebase . API\'s',
     credential:  'https://www.linkedin.com/learning/certificates/643f05463ae529f24bd9ea66a6ead9a20469bdb875a9ddda048c698eda3ee7c1',
   ),
-
-
-
   CertificateModel(
     name: 'Firebase Cloud Firestore',
     organization: 'LinkedIn',
@@ -74,7 +64,6 @@ List<CertificateModel> certificateList = [
     skills: 'Flutter . Dart . Firebase . FireStore',
     credential:  'https://www.linkedin.com/learning/certificates/8f8be25531d2bcdbab1972482150277f9a239a13ba4d314c0574638bf28d07d2',
   ),
-
   CertificateModel(
     name: 'Android App Security',
     organization: 'LinkedIn',
@@ -82,8 +71,6 @@ List<CertificateModel> certificateList = [
     skills: 'Pentesting . Android App',
     credential:  'https://www.linkedin.com/learning/certificates/1c6581b35d06edfbd6275d09e84b068e813880bf7d217b703716962d7aca3518',
   ),
-
-
   CertificateModel(
     name: 'Foundations of Cybersecurity',
     organization: 'Grow with Google on Coursera',
@@ -91,17 +78,13 @@ List<CertificateModel> certificateList = [
     skills: 'Ethical Hacking . Linux . Cyber Security ',
     credential:  'https://coursera.org/share/67e5cb0dd7c478f1d7ec81079c3a40b8',
   ),
-
-
-
- CertificateModel(
+  CertificateModel(
     name: 'HTML, CSS, and Javascript for Web Developers',
     organization: 'JUL Coursera',
     date: 'JUL 2023',
     skills: 'HTML . CSS . JS . Web Development',
     credential:  'https://coursera.org/share/67e5cb0dd7c478f1d7ec81079c3a40b8',
   ),
-
   CertificateModel(
     name: 'Network Defense Essentials (NDE)',
     organization: 'EC-Council',
