+++ lib/view/intro/components/animated_texts_componenets.dart
@@ -6,18 +6,14 @@ import '../../../view model/responsive.dart';
 class AnimatedImageContainer extends StatefulWidget {
   const AnimatedImageContainer({Key? key, this.height = 300, this.width = 250})
       : super(key: key);
-
   final double? width;
   final double? height;
-
   @override
-  _AnimatedImageContainerState createState() => _AnimatedImageContainerState();
+  AnimatedImageContainerState createState() => AnimatedImageContainerState();
 }
-
-class _AnimatedImageContainerState extends State<AnimatedImageContainer>
+class AnimatedImageContainerState extends State<AnimatedImageContainer>
     with SingleTickerProviderStateMixin {
   late AnimationController _controller;
-
   @override
   void initState() {
     super.initState();
@@ -26,13 +22,11 @@ class _AnimatedImageContainerState extends State<AnimatedImageContainer>
       duration: const Duration(milliseconds: 1000),
     )..repeat(reverse: true); // Repeat the animation loop
   }
-
   @override
   void dispose() {
     _controller.dispose();
     super.dispose();
   }
-
   @override
   Widget build(BuildContext context) {
     return AnimatedBuilder(
@@ -40,7 +34,7 @@ class _AnimatedImageContainerState extends State<AnimatedImageContainer>
       builder: (context, child) {
         final value = _controller.value;
         return Transform.translate(
-          offset: Offset(0, 2* value), // Move the container up and down
+          offset: Offset(0, 2 * value), // Move the container up and down
           child: Container(
             height: widget.height!,
             width: widget.width!,
@@ -75,98 +69,17 @@ class _AnimatedImageContainerState extends State<AnimatedImageContainer>
                 height: Responsive.isLargeMobile(context)
                     ? MediaQuery.sizeOf(context).width * 0.2
                     : Responsive.isTablet(context)
-                    ? MediaQuery.sizeOf(context).width * 0.14
-                    : 200,
+                        ? MediaQuery.sizeOf(context).width * 0.14
+                        : 200,
                 width: Responsive.isLargeMobile(context)
                     ? MediaQuery.sizeOf(context).width * 0.2
                     : Responsive.isTablet(context)
-                    ? MediaQuery.sizeOf(context).width * 0.14
-                    : 200,
+                        ? MediaQuery.sizeOf(context).width * 0.14
+                        : 200,
                 fit: BoxFit.cover,
               ),
             ),
           ),
-        );
-      },
-    );
-  }
-}
-
-
-class MyPortfolioText extends StatelessWidget {
-  const MyPortfolioText({super.key, required this.start, required this.end});
-
-  final double start;
-  final double end;
-
-  @override
-  Widget build(BuildContext context) {
-    return Builder(builder: (context) {
-      return TweenAnimationBuilder(
-        tween: Tween(begin: start, end: end),
-        duration: const Duration(milliseconds: 200),
-        builder: (context, value, child) {
-          return Text('My Personal Portfolio',
-              style: Theme.of(context).textTheme.headlineLarge!.copyWith(
-                  color: Colors.white,
-                  fontWeight: FontWeight.w900,
-                  height: 0,
-                  fontSize: value));
-        },
-      );
-    });
-  }
-}
-
-class AnimatedSubtitleText extends StatelessWidget {
-  final double start;
-  final double end;
-  final String text;
-  final bool gradient;
-  const AnimatedSubtitleText(
-      {super.key, required this.start, required this.end, required this.text, this.gradient=false,});
-
-  @override
-  Widget build(BuildContext context) {
-    return TweenAnimationBuilder(
-      tween: Tween(begin: start, end: end),
-      duration: const Duration(milliseconds: 200),
-      builder: (context, value, child) {
-        return Text(
-          text,
-          style: Theme.of(context).textTheme.headlineLarge!.copyWith(
-              color: Colors.white,
-              fontWeight: FontWeight.w900,
-              shadows: gradient? [
-                Shadow(color: Colors.pink,offset: Offset(0, 2),blurRadius: 10),
-                 Shadow(color: Colors.pink,offset: Offset(0, -2),blurRadius: 10),
-              ] :[] ,
-              height: 0,
-              fontSize: value),
-        );
-      },
-    );
-  }
-}
-
-class AnimatedDescriptionText extends StatelessWidget {
-  const AnimatedDescriptionText(
-      {super.key, required this.start, required this.end});
-
-  final double start;
-  final double end;
-
-  @override
-  Widget build(BuildContext context) {
-    return TweenAnimationBuilder(
-      tween: Tween(begin: start, end: end),
-      duration: const Duration(milliseconds: 200),
-      builder: (context, value, child) {
-        return Text(
-          'I\'m capable of creating excellent mobile apps, handling${Responsive.isLargeMobile(context) ? '\n' : ''}every step from ${!Responsive.isLargeMobile(context) ? '\n' : ''}concept to deployment.',
-          maxLines: 2,
-          overflow: TextOverflow.ellipsis,
-          style: TextStyle(color: Colors.grey, wordSpacing: 2, fontSize: value),
         );
       },
     );
