+++ lib/view/intro/components/description_text.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+import '../../../view model/responsive.dart';
+
+class AnimatedDescriptionText extends StatelessWidget {
+  const AnimatedDescriptionText(
+      {super.key, required this.start, required this.end});
+  final double start;
+  final double end;
+  @override
+  Widget build(BuildContext context) {
+    return TweenAnimationBuilder(
+      tween: Tween(begin: start, end: end),
+      duration: const Duration(milliseconds: 200),
+      builder: (context, value, child) {
+        return Text(
+          'I\'m capable of creating excellent mobile apps, handling${Responsive.isLargeMobile(context) ? '\n' : ''}every step from ${!Responsive.isLargeMobile(context) ? '\n' : ''}concept to deployment.',
+          maxLines: 2,
+          overflow: TextOverflow.ellipsis,
+          style: TextStyle(color: Colors.grey, wordSpacing: 2, fontSize: value),
+        );
+      },
+    );
+  }
+}
