+++ lib/view/main/components/drawer/about.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_portfolio/res/constants.dart';
+import 'drawer_image.dart';
 class About extends StatelessWidget {
   const About({super.key});
   @override
@@ -11,29 +12,7 @@ class About extends StatelessWidget {
         child: Column(
           children: [
             const Spacer(flex: 2,),
-            Container(
-              height: 100,
-              width: 100,
-              padding: const EdgeInsets.all(defaultPadding/6),
-              decoration: BoxDecoration(
-               borderRadius: BorderRadius.circular(50),
-                gradient: LinearGradient(
-                  colors: [
-                    Colors.pink,
-                    Colors.blue.shade900,
-                  ],
-                ),
-                boxShadow: const [
-                  BoxShadow(color: Colors.pink,blurRadius: 10,offset: Offset(0, 2)),
-                  BoxShadow(color: Colors.blue,blurRadius: 10,offset: Offset(0, -2)),
-                ]
-              ),
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(50),
-                  child: Transform.rotate(
-                      angle: 0.1,
-                      child: Image.asset('assets/images/profile.png',fit: BoxFit.cover,))),
-              ),
+            const DrawerImage(),
             const Spacer(),
             Text('Hamad Anwar',style: Theme.of(context).textTheme.titleSmall,),
             const SizedBox(height: defaultPadding/4,),
