+++ lib/view/main/components/drawer/personal_info.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+import '../../../../res/constants.dart';
+import 'header_info.dart';
+
+class PersonalInfo extends StatelessWidget {
+  const PersonalInfo({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        SizedBox(height: defaultPadding/2,),
+        AreaInfoText(title: 'Contact', text: '03166657602'),
+        AreaInfoText(title: 'Email', text: 'rh676838@gmail.com'),
+        AreaInfoText(title: 'LinkedIn', text: '@hamad-anwar'),
+        AreaInfoText(title: 'Github', text: '@hamad-anwar'),
+        SizedBox(
+          height: defaultPadding,
+        ),
+        Text('Skills',style: TextStyle(color: Colors.white),),
+        SizedBox(
+          height: defaultPadding,
+        ),
+      ],
+    );
+  }
+}
