+++ lib/view/splash/splash_view.dart
@@ -1,22 +1,15 @@
-
 import 'dart:async';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_portfolio/res/constants.dart';
 import 'package:flutter_portfolio/view/home/home.dart';
 import 'package:flutter_portfolio/view/intro/components/animated_texts_componenets.dart';
 import 'package:flutter_portfolio/view/splash/componenets/animated_loading_text.dart';
-
 class SplashView extends StatefulWidget {
   const SplashView({super.key});
-
   @override
   State<SplashView> createState() => _SplashViewState();
 }
-
 class _SplashViewState extends State<SplashView> {
-
-
   @override
   void initState() {
     // TODO: implement initState
@@ -25,7 +18,6 @@ class _SplashViewState extends State<SplashView> {
       Navigator.pushReplacement(context, MaterialPageRoute(builder: (context) => const HomePage(),));
     });
   }
-
   @override
   Widget build(BuildContext context) {
     return const Scaffold(
